split into main and mcshub
