+++ pkg/mcshub/controller.go
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package mcscontroller
+package mcshub
 
 import (
 	"context"
